requirements.txt
maafw
json-with-comments